requirements.txt
langchain
faiss-cpu
openai
pypdf
tqdm
python-dotenv
streamlit
